[tools] Move NormalizedStringComparer into its own file.
