+++ tools/dotnet-linker/dotnet-linker.csproj
@@ -77,6 +77,9 @@
     <Compile Include="..\common\LinkMode.cs">
       <Link>external/tools/common/LinkMode.cs</Link>
     </Compile>
+    <Compile Include="..\common\NormalizedStringComparer.cs">
+      <Link>tools\common\NormalizedStringComparer.cs</Link>
+    </Compile>
     <Compile Include="..\common\SdkVersions.cs">
       <Link>external/tools/common/SdkVersions.cs</Link>
     </Compile>
